+++ src/main/java/function/coverage/comparison/SiteClean.java
@@ -217,7 +217,7 @@ public class SiteClean {
                 covDiff = MathManager.abs(caseAvg, ctrlAvg);
             }
 
-            sb.append(FormatManager.getDouble(covDiff)).append(",");
+            sb.append(FormatManager.getFloat(covDiff)).append(",");
             sb.append(geneSize);
             return sb.toString();
         } else {
